+++ packages/svelte/src/lib/components/index.ts
@@ -1,2 +1,3 @@
 export * from './avatar'
 export * from './factory'
+export * from './timer'