+++ packages/svelte/src/lib/components/timer/examples/countdown.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+  import { Timer } from '@ark-ui/svelte/timer'
+</script>
+
+<Timer.Root autoStart countdown startMs={60 * 60 * 500}>
+  <Timer.Area>
+    <Timer.Item type="days" />
+    <Timer.Separator>:</Timer.Separator>
+    <Timer.Item type="hours" />
+    <Timer.Separator>:</Timer.Separator>
+    <Timer.Item type="minutes" />
+    <Timer.Separator>:</Timer.Separator>
+    <Timer.Item type="seconds" />
+  </Timer.Area>
+</Timer.Root>