+++ packages/svelte/src/lib/components/timer/use-timer-context.ts
@@ -1,0 +1,8 @@
+import { createContext } from '$lib/utils/create-context'
+import type { UseTimerReturn } from './use-timer.svelte'
+
+export interface UseTimerContext extends UseTimerReturn {}
+
+export const [TimerProvider, useTimerContext] = createContext<UseTimerContext>({
+  key: 'TimerContext',
+})