+++ packages/svelte/CHANGELOG.md
@@ -11,6 +11,7 @@ description: All notable changes will be documented in this file.
 - Added `Ark` factory component for `asChild` prop.
 - Added `Environment` component.
 - Added `Collection` helpers.
+- Added `Timer` component.
 
 ## [0.1.0] - 2024-11-27